restructured class inheritance
